fix:embed show
