+++ packages/blocks/src/embed-block/style.css
@@ -11,8 +11,13 @@
   color: inherit;
   text-align: center;
   color: var(--affine-placeholder-color);
+  height: 24px;
+  display: none;
 }
-
 .affine-embed-wrapper-caption::placeholder {
   color: var(--affine-placeholder-color);
+}
+
+.affine-embed-wrapper .caption-show {
+  display: inline-block;
 }
