+++ packages/blocks/src/page-block/default/default-page-block.ts
@@ -1,7 +1,6 @@
 /// <reference types="vite/client" />
 import { LitElement, html, css, unsafeCSS } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
-import { styleMap } from 'lit/directives/style-map.js';
 import {
   Disposable,
   Signal,
@@ -17,43 +16,22 @@ import {
   hotkey,
   BlockChildrenContainer,
   SelectionPosition,
-  HOTKEYS,
   assertExists,
-  isPageTitle,
-  getSplicedTitle,
-  noop,
-  getModelByElement,
-  matchFlavours,
-  focusPreviousBlock,
-  getDefaultPageBlock,
-  getBlockElementByModel,
-  getContainerByModel,
-  getPreviousBlock,
-  focusNextBlock,
-  Point,
-  caretRangeFromPoint,
-  getStartModelBySelection,
   getCurrentRange,
   isMultiBlockRange,
   getModelsByRange,
 } from '../../__internal__';
 import { DefaultSelectionManager } from './selection-manager';
-import {
-  batchUpdateTextType,
-  bindCommonHotkey,
-  deleteModels,
-  handleBackspace,
-  handleBlockSelectionBatchDelete,
-  handleSelectAll,
-  removeCommonHotKey,
-  tryUpdateGroupSize,
-  updateTextType,
-} from '../utils';
+import { deleteModels, tryUpdateGroupSize } from '../utils';
 import style from './style.css';
-import { CaptionIcon, CopyIcon, DeleteIcon, DownloadIcon } from '../icons';
-import { downloadImage, focusCaption, copyImgToClip } from './utils';
-import { isAtLineEdge } from '../../__internal__/rich-text/rich-text-operations';
-import { toolTipStyle } from '../../components/tooltip';
+
+import {
+  EmbedEditingContainer,
+  EmbedSelectedRectsContainer,
+  FrameSelectionRect,
+  SelectedRectsContainer,
+} from './rects-container';
+import { bindHotkeys, removeHotkeys } from './utils';
 
 export interface EmbedEditingState {
   position: { x: number; y: number };
@@ -77,243 +55,6 @@ function focusTextEnd(input: HTMLInputElement) {
   input.value = current;
 }
 
-function FrameSelectionRect(rect: DOMRect | null) {
-  if (rect === null) return html``;
-
-  const style = {
-    left: rect.left + 'px',
-    top: rect.top + 'px',
-    width: rect.width + 'px',
-    height: rect.height + 'px',
-  };
-  return html`
-    <style>
-      .affine-page-frame-selection-rect {
-        position: absolute;
-        background: var(--affine-selected-color);
-        z-index: 1;
-        pointer-events: none;
-      }
-    </style>
-    <div
-      class="affine-page-frame-selection-rect"
-      style=${styleMap(style)}
-    ></div>
-  `;
-}
-
-function EmbedSelectedRectsContainer(
-  rects: { left: number; top: number; width: number; height: number }[]
-) {
-  return html`
-    <style>
-      .affine-page-selected-embed-rects-container > div {
-        position: fixed;
-        border: 3px solid var(--affine-primary-color);
-      }
-    </style>
-    <div class="affine-page-selected-embed-rects-container resizable">
-      ${rects.map(rect => {
-        const style = {
-          display: 'block',
-          left: rect.left + 'px',
-          top: rect.top + 'px',
-          width: rect.width + 'px',
-          height: rect.height + 'px',
-        };
-        return html`<div class="resizes" style=${styleMap(style)}>
-          <div class="resize top-left"></div>
-          <div class="resize top-right"></div>
-          <div class="resize bottom-left"></div>
-          <div class="resize bottom-right"></div>
-        </div>`;
-      })}
-    </div>
-  `;
-}
-
-function SelectedRectsContainer(rects: DOMRect[]) {
-  return html`
-    <style>
-      .affine-page-selected-rects-container > div {
-        position: fixed;
-        background: var(--affine-selected-color);
-        z-index: 1;
-        pointer-events: none;
-        border-radius: 5px;
-      }
-    </style>
-    <div class="affine-page-selected-rects-container">
-      ${rects.map(rect => {
-        const style = {
-          display: 'block',
-          left: rect.left + 'px',
-          top: rect.top + 'px',
-          width: rect.width + 'px',
-          height: rect.height + 'px',
-        };
-        return html`<div style=${styleMap(style)}></div>`;
-      })}
-    </div>
-  `;
-}
-
-function EmbedEditingContainer(
-  embedEditingState: EmbedEditingState | null,
-  signals: DefaultPageSignals
-) {
-  if (embedEditingState) {
-    const style = {
-      left: embedEditingState.position.x + 'px',
-      top: embedEditingState.position.y + 'px',
-    };
-    return html`
-      <style>
-        .affine-image-option-container > div {
-          position: fixed;
-          z-index: 1;
-        }
-        ${toolTipStyle}
-      </style>
-
-      <div class="affine-image-option-container">
-        <div style=${styleMap(style)} class="image-option">
-          <format-bar-button
-            class="has-tool-tip"
-            width="100%"
-            @click=${() => focusCaption(embedEditingState.model)}
-          >
-            ${CaptionIcon}
-            <tool-tip inert tip-position="right" role="tooltip"
-              >Caption</tool-tip
-            >
-          </format-bar-button>
-          <format-bar-button
-            class="has-tool-tip"
-            width="100%"
-            @click=${() => {
-              assertExists(embedEditingState.model.source);
-              downloadImage(embedEditingState.model.source);
-            }}
-          >
-            ${DownloadIcon}
-            <tool-tip inert tip-position="right" role="tooltip"
-              >Download</tool-tip
-            >
-          </format-bar-button>
-          <format-bar-button
-            class="has-tool-tip"
-            width="100%"
-            @click=${() => {
-              assertExists(embedEditingState.model.source);
-              copyImgToClip(embedEditingState.model.source);
-            }}
-          >
-            ${CopyIcon}
-            <tool-tip inert tip-position="right" role="tooltip">Copy</tool-tip>
-          </format-bar-button>
-          <format-bar-button
-            class="has-tool-tip"
-            width="100%"
-            @click="${() => {
-              embedEditingState.model.page.deleteBlock(embedEditingState.model);
-              signals.updateEmbedRects.emit([]);
-            }}"
-          >
-            ${DeleteIcon}
-            <tool-tip inert tip-position="right" role="tooltip"
-              >Delete</tool-tip
-            >
-          </format-bar-button>
-        </div>
-      </div>
-    `;
-  } else {
-    return html``;
-  }
-}
-
-function handleUp(
-  selection: DefaultSelectionManager,
-  signals: DefaultPageSignals,
-  e: KeyboardEvent
-) {
-  const nativeSelection = window.getSelection();
-  if (nativeSelection?.anchorNode) {
-    const model = getStartModelBySelection();
-    const activeContainer = getContainerByModel(model);
-    const activePreNodeModel = getPreviousBlock(activeContainer, model.id);
-    const editableContainer = getBlockElementByModel(model)?.querySelector(
-      '.ql-editor'
-    ) as HTMLElement;
-    const range = nativeSelection.getRangeAt(0);
-    const { height, left, top } = range.getBoundingClientRect();
-    const newRange = caretRangeFromPoint(left, top - height / 2);
-    if (
-      (!newRange || !editableContainer.contains(newRange.startContainer)) &&
-      !isAtLineEdge(range)
-    ) {
-      // FIXME: Then it will turn the input into the div
-      if (
-        activePreNodeModel &&
-        matchFlavours(activePreNodeModel, ['affine:group'])
-      ) {
-        (
-          document.querySelector(
-            '.affine-default-page-block-title'
-          ) as HTMLInputElement
-        ).focus();
-      } else {
-        focusPreviousBlock(model, new Point(left, top));
-      }
-    }
-  } else {
-    signals.updateSelectedRects.emit([]);
-    const { state } = selection;
-    const selectedModel = getModelByElement(state.selectedBlocks[0]);
-    const container = getContainerByModel(selectedModel);
-    const preNodeModel = getPreviousBlock(container, selectedModel.id);
-    assertExists(preNodeModel);
-    const page = getDefaultPageBlock(selectedModel);
-    e.preventDefault();
-    focusPreviousBlock(
-      selectedModel,
-      page.lastSelectionPosition instanceof Point
-        ? page.lastSelectionPosition
-        : 'end'
-    );
-  }
-}
-function handleDown(
-  selection: DefaultSelectionManager,
-  signals: DefaultPageSignals,
-  e: KeyboardEvent
-) {
-  if (!selection.state.selectedBlocks.length) {
-    const nativeSelection = window.getSelection();
-    const model = getStartModelBySelection();
-    assertExists(nativeSelection);
-    const range = nativeSelection.getRangeAt(0);
-    const { left, bottom } = range.getBoundingClientRect();
-    focusNextBlock(model, new Point(left, bottom));
-  } else {
-    signals.updateSelectedRects.emit([]);
-    const { state } = selection;
-    const selectedModel = getModelByElement(state.selectedBlocks[0]);
-    const container = getContainerByModel(selectedModel);
-    const preNodeModel = getPreviousBlock(container, selectedModel.id);
-    assertExists(preNodeModel);
-    const page = getDefaultPageBlock(selectedModel);
-    e.preventDefault();
-    focusNextBlock(
-      selectedModel,
-      page.lastSelectionPosition instanceof Point
-        ? page.lastSelectionPosition
-        : 'start'
-    );
-    return;
-  }
-}
 @customElement('default-page-block')
 export class DefaultPageBlockComponent extends LitElement implements BlockHost {
   static styles = css`
@@ -373,178 +114,6 @@ export class DefaultPageBlockComponent extends LitElement implements BlockHost {
   @query('.affine-default-page-block-title')
   private _title!: HTMLInputElement;
 
-  private _bindHotkeys() {
-    const { page } = this;
-    const {
-      BACKSPACE,
-      SELECT_ALL,
-      H1,
-      H2,
-      H3,
-      H4,
-      H5,
-      H6,
-      SHIFT_UP,
-      SHIFT_DOWN,
-      NUMBERED_LIST,
-      BULLETED,
-      TEXT,
-      UP,
-      DOWN,
-      LEFT,
-      RIGHT,
-    } = HOTKEYS;
-
-    bindCommonHotkey(page);
-    hotkey.addListener(BACKSPACE, e => {
-      const { state } = this.selection;
-      if (state.type === 'native') {
-        handleBackspace(page, e);
-        return;
-      } else if (state.type === 'block') {
-        const { selectedBlocks } = state;
-        handleBlockSelectionBatchDelete(
-          page,
-          selectedBlocks.map(block => getModelByElement(block))
-        );
-        e.preventDefault();
-        state.clear();
-        this.signals.updateSelectedRects.emit([]);
-        this.signals.updateEmbedRects.emit([]);
-        this.signals.updateEmbedEditingState.emit(null);
-        return;
-      }
-      if (isPageTitle(e)) {
-        const target = e.target as HTMLInputElement;
-        // range delete
-        if (target.selectionStart !== target.selectionEnd) {
-          e.preventDefault();
-          const title = getSplicedTitle(target);
-          page.updateBlock(this.model, { title });
-          page.workspace.setPageMeta(page.id, { title });
-        }
-        // collapsed delete
-        else {
-          noop();
-        }
-        return;
-      }
-    });
-
-    hotkey.addListener(SELECT_ALL, e => {
-      e.preventDefault();
-      handleSelectAll();
-      this.selection.state.type = 'native';
-    });
-
-    hotkey.addListener(UP, e => {
-      handleUp(this.selection, this.signals, e);
-    });
-    hotkey.addListener(DOWN, e => {
-      handleDown(this.selection, this.signals, e);
-    });
-    hotkey.addListener(LEFT, e => {
-      let model: BaseBlockModel | null = null;
-      if (this.selection.state.selectedBlocks.length) {
-        model = getModelByElement(this.selection.state.selectedBlocks[0]);
-        this.signals.updateSelectedRects.emit([]);
-        this.selection.state.clear();
-        e.preventDefault();
-      } else {
-        const range = window.getSelection()?.getRangeAt(0);
-        if (range && range.collapsed && range.startOffset === 0) {
-          model = getStartModelBySelection();
-        }
-      }
-      model && focusPreviousBlock(model, 'end');
-    });
-    hotkey.addListener(RIGHT, e => {
-      let model: BaseBlockModel | null = null;
-      if (this.selection.state.selectedBlocks.length) {
-        model = getModelByElement(
-          this.selection.state.selectedBlocks[
-            this.selection.state.selectedBlocks.length - 1
-          ]
-        );
-        this.signals.updateSelectedRects.emit([]);
-        this.selection.state.clear();
-        e.preventDefault();
-      } else {
-        const range = window.getSelection()?.getRangeAt(0);
-        const textModel = getStartModelBySelection();
-        if (
-          range &&
-          range.collapsed &&
-          range.startOffset === textModel.text?.length
-        ) {
-          // handleUp(this.selection, this.signals);
-          model = getStartModelBySelection();
-        }
-      }
-      model && focusNextBlock(model, 'start');
-    });
-
-    hotkey.addListener(H1, () =>
-      this._updateType('affine:paragraph', 'h1', page)
-    );
-    hotkey.addListener(H2, () =>
-      this._updateType('affine:paragraph', 'h2', page)
-    );
-    hotkey.addListener(H3, () =>
-      this._updateType('affine:paragraph', 'h3', page)
-    );
-    hotkey.addListener(H4, () =>
-      this._updateType('affine:paragraph', 'h4', page)
-    );
-    hotkey.addListener(H5, () =>
-      this._updateType('affine:paragraph', 'h5', page)
-    );
-    hotkey.addListener(H6, () =>
-      this._updateType('affine:paragraph', 'h6', page)
-    );
-    hotkey.addListener(NUMBERED_LIST, () =>
-      this._updateType('affine:list', 'numbered', page)
-    );
-    hotkey.addListener(BULLETED, () =>
-      this._updateType('affine:list', 'bulleted', page)
-    );
-    hotkey.addListener(TEXT, () =>
-      this._updateType('affine:paragraph', 'text', page)
-    );
-    hotkey.addListener(SHIFT_UP, e => {
-      // TODO expand selection up
-    });
-    hotkey.addListener(SHIFT_DOWN, e => {
-      // TODO expand selection down
-    });
-
-    // !!!
-    // Don't forget to remove hotkeys at `_removeHotkeys`
-  }
-
-  private _removeHotkeys() {
-    removeCommonHotKey();
-    hotkey.removeListener([
-      HOTKEYS.BACKSPACE,
-      HOTKEYS.SELECT_ALL,
-      HOTKEYS.H1,
-      HOTKEYS.H2,
-      HOTKEYS.H3,
-      HOTKEYS.H4,
-      HOTKEYS.H5,
-      HOTKEYS.H6,
-      HOTKEYS.SHIFT_UP,
-      HOTKEYS.SHIFT_DOWN,
-      HOTKEYS.BULLETED,
-      HOTKEYS.NUMBERED_LIST,
-      HOTKEYS.TEXT,
-      HOTKEYS.UP,
-      HOTKEYS.DOWN,
-      HOTKEYS.LEFT,
-      HOTKEYS.RIGHT,
-    ]);
-  }
-
   private _onTitleKeyDown(e: KeyboardEvent) {
     const hasContent = !this.page.isEmpty;
     const { page, model, _title } = this;
@@ -584,20 +153,6 @@ export class DefaultPageBlockComponent extends LitElement implements BlockHost {
     const title = (e.target as HTMLInputElement).value;
     page.updateBlock(this.model, { title });
     page.workspace.setPageMeta(page.id, { title });
-  }
-
-  private _updateType(flavour: string, type: string, page: Page) {
-    const { state } = this.selection;
-    if (state.selectedBlocks.length > 0) {
-      batchUpdateTextType(
-        flavour,
-        page,
-        state.selectedBlocks.map(block => getModelByElement(block)),
-        type
-      );
-    } else {
-      updateTextType(flavour, type, page);
-    }
   }
 
   private _clearSelection() {
@@ -653,7 +208,7 @@ export class DefaultPageBlockComponent extends LitElement implements BlockHost {
   };
 
   firstUpdated() {
-    this._bindHotkeys();
+    bindHotkeys(this.page, this.selection, this.signals, this.model);
 
     hotkey.enableHotkey();
     this.model.propsUpdated.on(() => {
@@ -709,7 +264,7 @@ export class DefaultPageBlockComponent extends LitElement implements BlockHost {
   override disconnectedCallback() {
     super.disconnectedCallback();
 
-    this._removeHotkeys();
+    removeHotkeys();
     this._scrollDisposable.dispose();
     this.selection.dispose();
     window.removeEventListener(
